Fixed skid pad lap direction and added lap fail button
Fixed laps showing left and right mirrored in complete laps on webpage.

Added a lap fail button that fails the last recorded lap and resets the current lap in skid pad. Failed laps do not count into statistics such as average and delta times.

Turned the ESP into AP_STA mode on WiFi so that it can connect to router while being an AP.

Lastly recorded the error I got when I clicked the reset button and the ESP crashed.